Board rows cols mistake save point
Multiple functions have rows and cols mixed up requiring re-doing much of the code. This is a save point before changes are made and inevitably breaking everything. Currently rows are treated as cols for some functions which causes index out of range errors if window height is not equal to window width.