Modified board
The board used to contain ASCII characters to represent RGB values and if the tile was walkable or not. This disallowed for the ability of a tile to consist of multiple things, like a building atop of a specific kind of tile as well as tracking the variables associated with what it may contain. Now the board contains tile objects which consist of terrain, structure, and critter objects. This will allow for a bit more depth moving forward and makes things like units inside of a structure atop a certain terrain, all with their own stats possible.

Currently the program populates the screen with these tile objects with random RGB values producing clown vomit.

Files such as  mover.py, board.py, and tiles.py have been removed.

critters.py, structures.py, and tile.py have been created and will be the home for their respective class.