Recreated critter wandering with new tile system
The old board system was just a long string of characters. This disallowed a tile to contain multiple objects. The new board system consist of tile objects which each contain a terrain, critter, and structure object. The dynamic spawning and random wandering from the old system has been re-created with the new system. Movement directions increased from 4 to 8. (Diagonal movements)